pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "indextts-mlx"
version = "0.1.0"
description = "IndexTTS-2 TTS inference on Apple Silicon via MLX"
requires-python = ">=3.10,<3.14"
dependencies = [
    "mlx>=0.21",
    "mlx-lm>=0.20",
    "numpy",
    "soundfile",
    "librosa",
    "sentencepiece",
    "click",
    "ftfy",
    "isbnlib",
    "pydub",
    "ebooklib",
    "beautifulsoup4",
    "lxml",
    "flask>=3.0",
    "pyloudnorm",
    "mutagen",
    "pyyaml",
    "apscheduler>=3.10,<4",
    "setuptools>=68,<71",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-timeout", "black"]
long = [
    "spacy>=3.7",
    "pysbd",
    # nemo_text_processing + pynini must be installed manually on macOS.
    # Requires Python 3.11 or 3.12. See README.md for instructions.
]
parity = [
    "torch",
    "torchaudio",
    "transformers",
    "safetensors",
    "huggingface_hub",
    "omegaconf",
]
convert = [
    "torch",
    "safetensors",
    "huggingface_hub",
]
srv = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "httpx>=0.24",
    "mlx-audio>=0.3",
]

[project.scripts]
indextts = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["indextts_mlx*", "cli*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.pytest.ini_options]
markers = [
    "parity: parity tests against reference implementations (skipped if deps missing)",
    "llm: tests that load the real LLM (skipped by default; run with --llm)",
    "srv_integration: srv integration tests that load real models (skipped by default; run with --srv-integration)",
]
